Refactor event form and add styling to event item
