+++ sass/pages/_events.scss
@@ -29,4 +29,16 @@
       }
     }
   }
+
+  .event-item {
+    .panel-title {
+      font-size: 2.5em;
+    }
+
+    .panel-content {
+      h3 {
+        font-size: 1.5em;
+      }
+    }
+  }
 }
