+++ views/events/event_card.jsx
