+++ views/events/event_feed.jsx
@@ -1,6 +1,6 @@
 var React = require('react');
 var LoungeWrapper = require('../layout.jsx');
-var EventPanel = require('./event_panel.jsx');
+var EventCard = require('./event_card.jsx');
 
 var EventFeed = React.createClass({
   events: [
@@ -10,7 +10,7 @@ var EventFeed = React.createClass({
   getPanels: function() {
     return this.events.map(function(event, i) {
       return (
-        <EventPanel event={event} key={i} />
+        <EventCard event={event} key={i} />
       );
     });
   },
@@ -22,13 +22,15 @@ var EventFeed = React.createClass({
             <div className="panel new-event">
               <div className="panel-body">
                 <h2 className="panel-title">Create a new event</h2>
-                <h3 className="panel-label">What's your event called?</h3>
-                <input type="text" className="form-control" placeholder="My awesome event name" />
-                <h3 className="panel-label">What's your event about?</h3>
-                <textarea type="text" className="form-control" placeholder="My awesome event description" />
-                <h3 className="panel-label">When's your event?</h3>
-                <input type="date" className="form-control" />
-                <button type="button" className="btn btn-success">Create Event</button>
+                <form>
+                  <h3 className="panel-label">What's your event called?</h3>
+                  <input type="text" className="form-control" placeholder="My awesome event name" />
+                  <h3 className="panel-label">What's your event about?</h3>
+                  <textarea type="text" className="form-control" placeholder="My awesome event description" />
+                  <h3 className="panel-label">When's your event?</h3>
+                  <input type="date" className="form-control" />
+                  <input type="submit" className="btn btn-success" value="Create Event" />
+                </form>
               </div>
             </div>
           </div>
